html/semantics/forms/the-button-element/button-setcustomvalidity.html

<!DOCTYPE HTML>
<title>button setCustomValidity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<button id='test'></button>

<script>

test(() => {
  let button = document.getElementById("test");
  assert_false(button.validity.customError);
  button.setCustomValidity("custom error");
  assert_true(button.validity.customError);
}, "button setCustomValidity is correct")

</script>
